+++ src/ts/Ads/Utilities/ProgrammaticTrackingService.ts
@@ -38,7 +38,9 @@ export enum AdmobMetric {
     AdmobOMEnabled = 'admob_om_enabled',
     AdmobOMInjected = 'admob_om_injected',
     AdmobOMSessionStart = 'admob_om_session_start',
-    AdmobOMSessionFinish = 'admob_om_session_finish'
+    AdmobOMSessionFinish = 'admob_om_session_finish',
+    AdmobOMImpression = 'admob_om_impression',
+    AdmobOMRegisteredImpression = 'admob_om_registered_impression'
 }
 
 export enum BannerMetric {